metricsconfig: Split the label filter logic into a separate file
This is a non-functional change, only reorganising code for easier development.
